fix: Undefined method `escape' for URI:Module
The gem paperclip stopped being supported and updated in 2018.
Christmas 2020 ruby 3.0 was released in which the URL:Module no
longer has an escape method. That has been deprecated in favour
of CGI.escape.

The paperclip gem has been forked to kt-paperclip. It is still
being actively maintained and supports ruby 3.0
Ref: https://rubygems.org/gems/kt-paperclip
Ref: https://github.com/thoughtbot/paperclip
Ref: https://github.com/kreeti/kt-paperclip
Ref: https://github.com/kreeti/kt-paperclip/commit/7d1f358632ccfd370d57b3a2c0e39c42486b0fa9

Switch to kt-paperclip. The alternative is to rework the
code to use ActiveStorage which now has equivalent functionality